pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "flask-dance>=7.1.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.31.0",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "numpy>=2.3.2",
    "openai>=1.100.2",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "scikit-learn>=1.7.1",
    "selenium>=4.35.0",
    "sift-stack-py>=0.8.4",
    "sqlalchemy>=2.0.43",
    "trafilatura>=2.0.0",
    "uvicorn>=0.35.0",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
]
